Add an experimental IROp to bytecode compiler
This introduces an alternative backend for IROp: instead of generating quotes
and assembling them with scala3-staging, we directly generate bytecode using
the new convenient [Java Classfile API](https://openjdk.org/jeps/8280389).

Even with this API, generating bytecode by hand is much more unwieldy than
manipulating quotes, but it is significantly faster: on the Bench_ci benchmarks
on my machine I see:

    ## Existing compiler
    Result "datalog.benchmarks.Bench_ci.compiled_default_unordered__":
      1,766 ±(99.9%) 0,201 s/op [Average]
    ## Existing interpreter
    Result "datalog.benchmarks.Bench_ci.interpreted_default_unordered__ci":
      0,554 ±(99.9%) 0,204 s/op [Average]
    ## New backend
    Result "datalog.benchmarks.Bench_ci.bytecodeGenerated__":
      0,587 ±(99.9%) 0,154 s/op [Average]

In other words, bytecode generation is within striking distance of the
interpreter and much faster than staging.

This implementation is a mostly mechanical translation of
`compileIR`/`compileIRRelOp`, it would be really interesting to investigate
whether the process of converting quotes and splices into bytecode generation
logic could be automated, maybe via a macro annotation.

Remaining TODOs:
- Support CompiledFnIndexed generation?

Note that because the Classfile API is not yet available in the JDK we
currently depend on a copy of the API from https://github.com/Glavo/classfile