Optimization: switch to representing values as Int instead of String
The Ints need to be boxed, but there's no unboxing/boxing happening in the core
of the program, so this isn't a big issue. And comparing Ints is much faster
than comparing String. This leads to a 40% speed-up on
datalog.benchmarks.examples.cspa10k_optimized.jit_indexed_sel__0_blocking_DELTA_lambda_EOL
on my laptop.