Refactor parallelism logic to make it easier to experiment
Enabling inParallel in other ops doesn't improve performance on
datalog.benchmarks.examples.cspa10k_optimized.interpreted_indexed_sel__0____EOL
on my laptop and in fact can make it worse. This needs to be tested more
thoroughly on a bigger dataset on a machine where all cores are the same (my
laptop has a mix of performance and efficiency cores which will likely add a lot
of noise to any benchmark result).